+++ src/main/java/io/github/astrapi69/collections/properties/IndexSortedProperties.java
@@ -31,9 +31,9 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 
-import io.github.astrapi69.comparators.NullCheckComparator;
 import io.github.astrapi69.collections.CollectionExtensions;
 import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.comparators.NullCheckComparator;
 
 /**
  * The class {@link IndexSortedProperties} extends SortedProperties and holds an intern list with