+++ src/main/java/io/github/astrapi69/comparators/LocaleComparator.java
@@ -27,8 +27,6 @@ package io.github.astrapi69.comparators;
 import java.util.Comparator;
 import java.util.Locale;
 
-import io.github.astrapi69.comparators.NullCheckComparator;
-
 /**
  * The class {@link LocaleComparator} compares {@linkplain Locale} objects. This Comparator does not
  * check null, for null check you can decorate it with the